+++ Wn32/Code/Plat/Gfx/nx/sprite.cpp
@@ -453,12 +453,12 @@ void sSprite::Draw( void )
 	});
 
 	// Push indices
-	m_indices.push_back(vi + 0);
-	m_indices.push_back(vi + 1);
-	m_indices.push_back(vi + 2);
-	m_indices.push_back(vi + 2);
-	m_indices.push_back(vi + 1);
-	m_indices.push_back(vi + 3);
+	m_indices.push_back((GLushort)(vi + 0));
+	m_indices.push_back((GLushort)(vi + 1));
+	m_indices.push_back((GLushort)(vi + 2));
+	m_indices.push_back((GLushort)(vi + 2));
+	m_indices.push_back((GLushort)(vi + 1));
+	m_indices.push_back((GLushort)(vi + 3));
 }
